requirements.txt
numpy==1.26.4
sounddevice==0.4.6
torch==2.0.0
tensorflow==2.15.0
openai-whisper==20231117
pydub==0.25.1
TTS==0.17.6
python-dotenv==1.0.0
ffmpeg-python==0.2.0
webrtcvad==2.0.10
openwakeword==0.6.0

# Optional dependencies for future expansion
# fastapi>=0.100.0
# chromadb>=0.4.0
# openai>=1.0.0
# boto3>=1.28.0
# pinecone-client>=2.2.0  # For vector database
# anthropic>=0.3.0  # For Claude API 